utils/requirements.txt
boto3==1.28.38
botocore==1.31.38
tabulate==0.9.0
python-dotenv==1.0.0